cfg: convert Build section to Tasks section
This is a first step to implement definition of multiple tasks instead of
supporting only to define a single build task.

The build section in the app config is converted to a list of tasks.
Tasks have additional a name.
To stay compatible with the remaining code, the tasks list must have for now
exactly 1 task that is named "build".