cfg: refactor Include config
Include config has now separate sections for Input, Outputs and Task
definitions.
The includes directive in the top-level app config can include Tasks.
The includes directive in the Tasks section can include inputs and outputs.

Handling of validation error was refactored. A ValidationError type is
introduced. The path to the element that failed validation is recorded in the
ValidationError.